Move AggregationJobDriver to the library target
